Basic hashing code
Added hashing code for frames so that they can be hashed.
Hashing is performed by capturing the Frame type with a local type
HashFrame which has a manual implementation of the Hash trait.
Hashing is done on a per pixel basis where each pixel of all three
planes of a given frame is iterated over and hashed. This is not an
optimal method of hashing the frames since it will be very intensive but
will be acceptable for a proof of concept.
This method produces hashes which are, predicatably, the same for
identical frames but very different for nonidential frames.
This means the hashing method is working as intended and can be used to
begin implementing the hashmap method of frame storage.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>