Fixed collisions
Added additional information to the hash's entropy pool to stop
collisions from occuring. Will need to go back and tune this later to
minimize the amount of data which needs to be hashed to ensure the
minimum work is needed to generate a hash for a given tile's
coefficients.
Note: It may be the case that the additional parameters are negligable
in comparision to the coefficients themselves.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>