Added encoding with hashes
Coefficents are added to a hash table when they are seen for the first
time. When they are seen again, a marker bit is set to 1 in the
bitstream and the hash is sent before the eob skip bit is written. Then
the coefficients are spoofed to be all zero and are written accordingly.
This breaks compatability with a standard AV1 decoder.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>