Updated hash read and write method
Now instead of one global hashtable, mutiple are used. One for looking
up hashes to use, and one to store new hashes. The new hashes are stored
in a queue of vectors holding all the new hashes found that frame. At
the end of each frame, the oldest of these vectors (from 5 frames ago)
is read into the global lookup table (now implemented as a RwLock) which
allows further frames to be encoded using these hashes. Then a new vec
is added to the end of the list. When a new hash is found it is added to
the last vec to ensure 5 full frames pass before it is used. This is to
ensure that the decoder is able to read the corosponding tile which the
hash relates to and add it to its hashmap before it is used to decode a
tile.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>