Reduce information of hash
Since the goal is to maximize collisions we can throw out a bunch of
information we were previously using to hash the objects.
Specifically, we can throw away everything except the coefficients to
further increase the likelyhood we actually collide with hashes.
Seems to reduce the size of the resulting video significantly.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>