Possible fix for hash writing bug
Problem may have been that we were considering hashes which we found
during trial encoding passes.
This is bad since we might not actually use these passes.
So we either refuse to use hashes found during these passes (pass None)
or we can simply discard the hashes we found later when we decide we
don't want to use that encoding pass.